clean python files
