+++ dags/subs.py
@@ -1,4 +1,6 @@
 from math import radians, cos, sin, asin, sqrt
+
+
 def haversine(lon1, lat1, lon2, lat2):
     """
     Calculate the great circle distance in kilometers between two points
@@ -15,8 +17,10 @@ def haversine(lon1, lat1, lon2, lat2):
     dlat = lat2 - lat1
     a = sin(dlat / 2) ** 2 + cos(lat1) * cos(lat2) * sin(dlon / 2) ** 2
     c = 2 * asin(sqrt(a))
-    r = 6371  # Radius of earth in kilometers. Use 3956 for miles. Determines return value units.
+    # Radius of earth in kilometers. Use 3956 for miles. Determines return value units.
+    r = 6371
     return c * r
+
 
 '''lon1=-71.064544
 lat1=42.28787
